Extract Predictor Names from Formula or Terms

`all.vars` returns all variables used in a formula. This function only returns the variables explicitly used on the right-hand side (i.e., it will not resolve dots unless the object is terms with a data set specified).

form_pred(object, ...)

Arguments

object
A model formula or [stats::terms()] object.

...
Arguments to pass to [all.vars()]

Value

A character vector of names

Examples

form_pred(y ~ x + z)

#> [1] "x" "z"

form_pred(terms(y ~ x + z))

#> [1] "x" "z"

form_pred(y ~ x + log(z))

#> [1] "x" "z"

form_pred(log(y) ~ x + z)

#> [1] "x" "z"

form_pred(y1 + y2 ~ x + z)

#> [1] "x" "z"

form_pred(log(y1) + y2 ~ x + z)

#> [1] "x" "z"

# will fail:
# form_pred(y ~ .)

form_pred(terms(Species ~ (.)^2, data = iris))

#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width" 

form_pred(terms( ~ (.)^2, data = iris))

#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"