Leave-One-Out Cross-Validation

Leave-one-out (LOO) cross-validation uses one data point in the original set as the assessment data and all other data points as the analysis set. A LOO resampling set has ass many resamples as rows in the original data set.

loo_cv(data, ...)

Arguments

data
A data frame.

...
Not currently used.

Value

An tibble with classes loo_cv, rset, tbl_df, tbl, and data.frame. The results include a column for the data split objects and one column called id that has a character string with the resample identifier.

Examples

loo_cv(mtcars)

#> # Leave-one-out cross-validation
#> # A tibble: 32 x 2
#>          splits         id
#>          <list>      <chr>
#>  1 <S3: rsplit>  Resample1
#>  2 <S3: rsplit>  Resample2
#>  3 <S3: rsplit>  Resample3
#>  4 <S3: rsplit>  Resample4
#>  5 <S3: rsplit>  Resample5
#>  6 <S3: rsplit>  Resample6
#>  7 <S3: rsplit>  Resample7
#>  8 <S3: rsplit>  Resample8
#>  9 <S3: rsplit>  Resample9
#> 10 <S3: rsplit> Resample10
#> # ... with 22 more rows