Create or Modify Stratification Variables

For stratified resampling, this function can create strata from numeric data and also make non-numeric data more conducive to be used for stratification.

make_strata(x, nunique = 5, pool = 0.15, depth = 20)

Arguments

x
An input vector.

nunique
An integer for the number of unique value threshold in the algorithm.

pool
A proportion of data used to determine if a particular group is too small and should be pooled into another group.

depth
An integer that is used to determine the best number of percentiles that should be used. The number of bins are based on `min(5, floor(n / depth))` where `n = length(x)`. If `x` is numeric, there must be at least 40 rows in the data set (when `depth = 20`) to conduct stratified sampling.

Value

A factor vector.

Details

For numeric data, if the number of unique levels is less than `nunique`, the data are treated as categorical data.

For categorical inputs, the function will find levels of `x` than occur in the data with percentage less than `pool`. The values from these groups will be randomly assigned to the remaining strata (as will data points that have missing values in `x`). For numeric data with more unique values than `nunique`, the data will be converted to being categorical based on percentiles of the data. The percentile groups will have no more than 20 percent of the data in each group. Again, missing values in `x` are randomly assigned to groups.

Examples

set.seed(61)
x1 <- rpois(100, lambda = 5)
table(x1)

#> x1
#>  1  2  3  4  5  6  7  8  9 10 11 
#>  3 16  8 19 14 18 11  4  5  1  1 

table(make_strata(x1))

#> 
#>   [1,3]   (3,4] (4,5.4] (5.4,7]  (7,11] 
#>      27      19      14      29      11 

set.seed(554)
x2 <- rpois(100, lambda = 1)
table(x2)

#> x2
#>  0  1  2  3  4 
#> 36 34 19  6  5 

table(make_strata(x2))

#> 
#>  0  1  2 
#> 36 38 26 

# small groups are randomly assigned
x3 <- factor(x2)
table(x3)

#> x3
#>  0  1  2  3  4 
#> 36 34 19  6  5 

table(make_strata(x3))

#> 
#>  0  1  2 
#> 41 37 22 

# `oilType` data from `caret`
x4 <- rep(LETTERS[1:7], c(37, 26, 3, 7, 11, 10, 2))
table(x4)

#> x4
#>  A  B  C  D  E  F  G 
#> 37 26  3  7 11 10  2 

table(make_strata(x4))

#> 
#>  A  B 
#> 54 42 

table(make_strata(x4, pool = 0.1))

#> 
#>  A  B  E  F 
#> 39 29 13 15 

table(make_strata(x4, pool = 0.0))

#> 
#>  A  B  C  D  E  F  G 
#> 37 26  3  7 11 10  2 

# not enough data to stratify
x5 <- rnorm(20)
table(make_strata(x5))

#> Warning: Too little data to stratify. Unstratified resampling will be used.

#> 
#> strata1 
#>      20 

set.seed(483)
x6 <- rnorm(200)
quantile(x6, probs = (0:5)/5)

#>         0%        20%        40%        60%        80%       100% 
#> -2.9114060 -0.9513821 -0.3286468  0.2027140  0.7050643  2.6792505 

table(make_strata(x6))

#> 
#>  [-2.91,-0.951] (-0.951,-0.329]  (-0.329,0.203]   (0.203,0.705]    (0.705,2.68] 
#>              40              40              40              40              40