Monte Carlo Cross-Validation

One resample of Monte Carlo cross-validation takes a random sample (without replacement) of the original data set to be used for analysis. All other data points are added to the assessment set.

mc_cv(data, prop = 3/4, times = 25, strata = NULL, ...)

Arguments

data
A data frame.

prop
The proportion of data to be retained for modeling/analysis.

times
The number of times to repeat the sampling..

strata
A variable that is used to conduct stratified sampling to create the resamples. (not working yet)

...
Not currently used.

Value

An tibble with classes mc_cv, rset, tbl_df, tbl, and data.frame. The results include a column for the data split objects and a column called id that has a character string with the resample identifier.

Details

The strata argument causes the random sampling to be conducted within the stratification variable. The can help ensure that the number of data points in the analysis data is equivalent to the proportions in the original data set.

Examples

mc_cv(mtcars, times = 2)

#> # Monte Carlo cross-validation (0.75/0.25) with 2 resamples 
#> # A tibble: 2 x 2
#>         splits        id
#>         <list>     <chr>
#> 1 <S3: rsplit> Resample1
#> 2 <S3: rsplit> Resample2

mc_cv(mtcars, prop = .5, times = 2)

#> # Monte Carlo cross-validation (0.5/0.5) with 2 resamples 
#> # A tibble: 2 x 2
#>         splits        id
#>         <list>     <chr>
#> 1 <S3: rsplit> Resample1
#> 2 <S3: rsplit> Resample2