Job Attrition

Job Attrition

Source

The IBM Watson Analytics Lab website https://www.ibm.com/communities/analytics/watson-analytics-blog/hr-employee-attrition/

Value

attrition

a data frame

Details

These data are from the IBM Watson Analytics Lab. The website describes the data with “Uncover the factors that lead to employee attrition and explore important questions such as ‘show me a breakdown of distance from home by job role and attrition’ or ‘compare average monthly income by education and attrition’. This is a fictional data set created by IBM data scientists.”. There are 1470 rows.

Examples

data(attrition)
str(attrition)

#> 'data.frame':	1470 obs. of  31 variables:
#>  $ Age                     : int  41 49 37 33 27 32 59 30 38 36 ...
#>  $ Attrition               : Factor w/ 2 levels "No","Yes": 2 1 2 1 1 1 1 1 1 1 ...
#>  $ BusinessTravel          : Factor w/ 3 levels "Non-Travel","Travel_Frequently",..: 3 2 3 2 3 2 3 3 2 3 ...
#>  $ DailyRate               : int  1102 279 1373 1392 591 1005 1324 1358 216 1299 ...
#>  $ Department              : Factor w/ 3 levels "Human_Resources",..: 3 2 2 2 2 2 2 2 2 2 ...
#>  $ DistanceFromHome        : int  1 8 2 3 2 2 3 24 23 27 ...
#>  $ Education               : Ord.factor w/ 5 levels "Below_College"<..: 2 1 2 4 1 2 3 1 3 3 ...
#>  $ EducationField          : Factor w/ 6 levels "Human_Resources",..: 2 2 5 2 4 2 4 2 2 4 ...
#>  $ EnvironmentSatisfaction : Ord.factor w/ 4 levels "Low"<"Medium"<..: 2 3 4 4 1 4 3 4 4 3 ...
#>  $ Gender                  : Factor w/ 2 levels "Female","Male": 1 2 2 1 2 2 1 2 2 2 ...
#>  $ HourlyRate              : int  94 61 92 56 40 79 81 67 44 94 ...
#>  $ JobInvolvement          : Ord.factor w/ 4 levels "Low"<"Medium"<..: 3 2 2 3 3 3 4 3 2 3 ...
#>  $ JobLevel                : int  2 2 1 1 1 1 1 1 3 2 ...
#>  $ JobRole                 : Factor w/ 9 levels "Healthcare_Representative",..: 8 7 3 7 3 3 3 3 5 1 ...
#>  $ JobSatisfaction         : Ord.factor w/ 4 levels "Low"<"Medium"<..: 4 2 3 3 2 4 1 3 3 3 ...
#>  $ MaritalStatus           : Factor w/ 3 levels "Divorced","Married",..: 3 2 3 2 2 3 2 1 3 2 ...
#>  $ MonthlyIncome           : int  5993 5130 2090 2909 3468 3068 2670 2693 9526 5237 ...
#>  $ MonthlyRate             : int  19479 24907 2396 23159 16632 11864 9964 13335 8787 16577 ...
#>  $ NumCompaniesWorked      : int  8 1 6 1 9 0 4 1 0 6 ...
#>  $ OverTime                : Factor w/ 2 levels "No","Yes": 2 1 2 2 1 1 2 1 1 1 ...
#>  $ PercentSalaryHike       : int  11 23 15 11 12 13 20 22 21 13 ...
#>  $ PerformanceRating       : Ord.factor w/ 4 levels "Low"<"Good"<"Excellent"<..: 3 4 3 3 3 3 4 4 4 3 ...
#>  $ RelationshipSatisfaction: Ord.factor w/ 4 levels "Low"<"Medium"<..: 1 4 2 3 4 3 1 2 2 2 ...
#>  $ StockOptionLevel        : int  0 1 0 0 1 0 3 1 0 2 ...
#>  $ TotalWorkingYears       : int  8 10 7 8 6 8 12 1 10 17 ...
#>  $ TrainingTimesLastYear   : int  0 3 3 3 3 2 3 2 2 3 ...
#>  $ WorkLifeBalance         : Ord.factor w/ 4 levels "Bad"<"Good"<"Better"<..: 1 3 3 3 3 2 2 3 3 2 ...
#>  $ YearsAtCompany          : int  6 10 0 8 2 7 1 1 9 7 ...
#>  $ YearsInCurrentRole      : int  4 7 0 7 2 7 0 0 7 7 ...
#>  $ YearsSinceLastPromotion : int  0 1 0 3 2 3 0 0 1 7 ...
#>  $ YearsWithCurrManager    : int  5 7 0 0 2 6 0 0 8 7 ...