Gather an `rset` Object

This method uses `gather` on an `rset` object to stack all of the non-ID or split columns in the data and is useful for stacking model evaluation statistics. The resulting data frame has a column based on the column names of `data` and another for the values.

# S3 method for rset
gather(data, key = NULL, value = NULL, ..., na.rm = TRUE,
  convert = FALSE, factor_key = TRUE)

Arguments

data
An `rset` object.

key, value, ...
Not specified in this method and will be ignored. Note that this means that selectors are ignored if they are passed to the function.

na.rm
If `TRUE`, will remove rows from output where the value column in `NA`.

convert
If `TRUE` will automatically run `type.convert()` on the key column. This is useful if the column names are actually numeric, integer, or logical.

factor_key
If FALSE, the default, the key values will be stored as a character vector. If `TRUE`, will be stored as a factor, which preserves the original ordering of the columns.

Value

A data frame with the ID columns, a column called `model` (with the previous column names), and a column called `statistic` (with the values).

Examples

library(rsample)
cv_obj <- vfold_cv(mtcars, v = 10)
cv_obj$lm_rmse <- rnorm(10, mean = 2)
cv_obj$nnet_rmse <- rnorm(10, mean = 1)
gather(cv_obj)

#>        id     model   statistic
#> 1  Fold01   lm_rmse  3.01374433
#> 2  Fold02   lm_rmse  1.79623542
#> 3  Fold03   lm_rmse  3.14856675
#> 4  Fold04   lm_rmse  3.34679349
#> 5  Fold05   lm_rmse  2.72970634
#> 6  Fold06   lm_rmse  2.20950131
#> 7  Fold07   lm_rmse  3.12815485
#> 8  Fold08   lm_rmse  1.66607666
#> 9  Fold09   lm_rmse  1.51367095
#> 10 Fold10   lm_rmse  2.78138724
#> 11 Fold01 nnet_rmse  1.12502893
#> 12 Fold02 nnet_rmse -0.04019181
#> 13 Fold03 nnet_rmse  0.83509883
#> 14 Fold04 nnet_rmse  1.82513817
#> 15 Fold05 nnet_rmse  2.84551869
#> 16 Fold06 nnet_rmse  1.61097953
#> 17 Fold07 nnet_rmse  3.46298079
#> 18 Fold08 nnet_rmse  1.44740391
#> 19 Fold09 nnet_rmse  1.60834335
#> 20 Fold10 nnet_rmse  1.47751459