Simple Training/Test Set Splitting

initial_split creates a single binary split of the data into a training set and testing set. initial_time_split does the same, but takes the first prop samples for training, instead of a random selection. training and testing are used to extract the resulting data.

initial_split(data, prop = 3/4, strata = NULL, breaks = 4, ...)

initial_time_split(data, prop = 3/4, ...)

training(x)

testing(x)

Arguments

data
A data frame.

prop
The proportion of data to be retained for modeling/analysis.

strata
A variable that is used to conduct stratified sampling to create the resamples. This could be a single character value or a variable name that corresponds to a variable that exists in the data frame.

breaks
A single number giving the number of bins desired to stratify a numeric stratification variable.

...
Not currently used.

x
An rsplit object produced by initial_split

Value

An rset object that can be used with the training and testing functions to extract the data in each split.

Details

The strata argument causes the random sampling to be conducted within the stratification variable. The can help ensure that the number of data points in the training data is equivalent to the proportions in the original data set.

Examples

set.seed(1353)
car_split <- initial_split(mtcars)
train_data <- training(car_split)
test_data <- testing(car_split)

drinks_split <- initial_time_split(drinks)
train_data <- training(drinks_split)
test_data <- testing(car_split)
c(max(train_data$date), min(test_data$date))  # no overlap

#> Warning: no non-missing arguments to min; returning Inf

#> [1] "2011-03-01" NA