Monte Carlo Cross-Validation

One resample of Monte Carlo cross-validation takes a random sample (without replacement) of the original data set to be used for analysis. All other data points are added to the assessment set.

mc_cv(data, prop = 3/4, times = 25, strata = NULL, ...)

Arguments

data
A data frame.

prop
The proportion of data to be retained for modeling/analysis.

times
The number of times to repeat the sampling..

strata
A variable that is used to conduct stratified sampling to create the resamples.

...
Not currently used.

Value

An tibble with classes mc_cv, rset, tbl_df, tbl, and data.frame. The results include a column for the data split objects and a column called id that has a character string with the resample identifier.

Details

The strata argument causes the random sampling to be conducted within the stratification variable. The can help ensure that the number of data points in the analysis data is equivalent to the proportions in the original data set.

Examples

mc_cv(mtcars, times = 2)

#> # # Monte Carlo cross-validation (0.75/0.25) with 2 resamples  
#> [38;5;246m# A tibble: 2 x 2[39m
#>   splits         id       
#>   [3m[38;5;246m<list>[39m[23m         [3m[38;5;246m<chr>[39m[23m    
#> [38;5;250m1[39m [38;5;246m<split [24/8]>[39m Resample1
#> [38;5;250m2[39m [38;5;246m<split [24/8]>[39m Resample2

mc_cv(mtcars, prop = .5, times = 2)

#> # # Monte Carlo cross-validation (0.5/0.5) with 2 resamples  
#> [38;5;246m# A tibble: 2 x 2[39m
#>   splits          id       
#>   [3m[38;5;246m<list>[39m[23m          [3m[38;5;246m<chr>[39m[23m    
#> [38;5;250m1[39m [38;5;246m<split [16/16]>[39m Resample1
#> [38;5;250m2[39m [38;5;246m<split [16/16]>[39m Resample2

library(purrr)
iris2 <- iris[1:130, ]

set.seed(13)
resample1 <- mc_cv(iris2, times = 3, prop = .5)
map_dbl(resample1$splits,
        function(x) {
          dat <- as.data.frame(x)$Species
          mean(dat == "virginica")
        })

#> [1] 0.2769231 0.2615385 0.2461538

set.seed(13)
resample2 <- mc_cv(iris2, strata = "Species", times = 3, prop = .5)
map_dbl(resample2$splits,
        function(x) {
          dat <- as.data.frame(x)$Species
          mean(dat == "virginica")
        })

#>         1         2         3 
#> 0.2307692 0.2307692 0.2307692