Nested or Double Resampling

nested_cv can be used to take the results of one resampling procedure and conduct further resamples within each split. Any type of resampling used in rsample can be used.

nested_cv(data, outside, inside)

Arguments

data
A data frame.

outside
The initial resampling specification. This can be an already created object or an expression of a new object (see the examples below). If the latter is used, the data argument does not need to be specified and, if it is given, will be ignored.

inside
An expression for the type of resampling to be conducted within the initial procedure.

Value

An tibble with classe nested_cv and any other classes that outer resampling process normally contains. The results include a column for the outer data split objects, one or more id columns, and a column of nested tibbles called inner_resamples with the additional resamples.

Details

It is a bad idea to use bootstrapping as the outer resampling procedure (see the example below)

Examples

## Using expressions for the resampling procedures:
nested_cv(mtcars, outside = vfold_cv(v = 3), inside = bootstraps(times = 5))

#> [1] "nested_cv"  "vfold_cv"   "rset"       "tbl_df"     "tbl"       
#> [6] "data.frame"
#> # Nested resampling:
#> #  outer: 3-fold cross-validation
#> #  inner: Bootstrap sampling
#> [38;5;246m# A tibble: 3 x 3[39m
#>   splits          id    inner_resamples 
#>   [3m[38;5;246m<list>[39m[23m          [3m[38;5;246m<chr>[39m[23m [3m[38;5;246m<list>[39m[23m          
#> [38;5;250m1[39m [38;5;246m<split [21/11]>[39m Fold1 [38;5;246m<tibble [5 × 2]>[39m
#> [38;5;250m2[39m [38;5;246m<split [21/11]>[39m Fold2 [38;5;246m<tibble [5 × 2]>[39m
#> [38;5;250m3[39m [38;5;246m<split [22/10]>[39m Fold3 [38;5;246m<tibble [5 × 2]>[39m

## Using an existing object:
folds <- vfold_cv(mtcars)
nested_cv(mtcars, folds, inside = bootstraps(times = 5))

#> [1] "nested_cv"  "vfold_cv"   "rset"       "tbl_df"     "tbl"       
#> [6] "data.frame"
#> # Nested resampling:
#> #  outer: `folds`
#> #  inner: Bootstrap sampling
#> [38;5;246m# A tibble: 10 x 3[39m
#>    splits         id     inner_resamples 
#>    [3m[38;5;246m<list>[39m[23m         [3m[38;5;246m<chr>[39m[23m  [3m[38;5;246m<list>[39m[23m          
#> [38;5;250m 1[39m [38;5;246m<split [28/4]>[39m Fold01 [38;5;246m<tibble [5 × 2]>[39m
#> [38;5;250m 2[39m [38;5;246m<split [28/4]>[39m Fold02 [38;5;246m<tibble [5 × 2]>[39m
#> [38;5;250m 3[39m [38;5;246m<split [29/3]>[39m Fold03 [38;5;246m<tibble [5 × 2]>[39m
#> [38;5;250m 4[39m [38;5;246m<split [29/3]>[39m Fold04 [38;5;246m<tibble [5 × 2]>[39m
#> [38;5;250m 5[39m [38;5;246m<split [29/3]>[39m Fold05 [38;5;246m<tibble [5 × 2]>[39m
#> [38;5;250m 6[39m [38;5;246m<split [29/3]>[39m Fold06 [38;5;246m<tibble [5 × 2]>[39m
#> [38;5;250m 7[39m [38;5;246m<split [29/3]>[39m Fold07 [38;5;246m<tibble [5 × 2]>[39m
#> [38;5;250m 8[39m [38;5;246m<split [29/3]>[39m Fold08 [38;5;246m<tibble [5 × 2]>[39m
#> [38;5;250m 9[39m [38;5;246m<split [29/3]>[39m Fold09 [38;5;246m<tibble [5 × 2]>[39m
#> [38;5;250m10[39m [38;5;246m<split [29/3]>[39m Fold10 [38;5;246m<tibble [5 × 2]>[39m

## The dangers of outer bootstraps:
set.seed(2222)
bad_idea <- nested_cv(mtcars,
                      outside = bootstraps(times = 5),
                      inside = vfold_cv(v = 3))

#> Warning: Using bootstrapping as the outer resample is dangerous since the inner resample might have the same data point in both the analysis and assessment set.

first_outer_split <- bad_idea$splits[[1]]
outer_analysis <- as.data.frame(first_outer_split)
sum(grepl("Volvo 142E", rownames(outer_analysis)))

#> [1] 3

## For the 3-fold CV used inside of each bootstrap, how are the replicated
## `Volvo 142E` data partitioned?
first_inner_split <- bad_idea$inner_resamples[[1]]$splits[[1]]
inner_analysis <- as.data.frame(first_inner_split)
inner_assess   <- as.data.frame(first_inner_split, data = "assessment")

sum(grepl("Volvo 142E", rownames(inner_analysis)))

#> [1] 2

sum(grepl("Volvo 142E", rownames(inner_assess)))

#> [1] 1